+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # latest
-Using version [2.0](#20-2025-12-30)
+Using version [2.1](#21-2026-01-01)
 
 # Major version 2
-Using version [2.0](#20-2025-12-30)
+Using version [2.1](#21-2026-01-01)
 
 # Major version 1
 Using version [1.6](#16-2025-12-18)
+
+## 2.1 (2026-01-01)
+**Changes**
+- Extract first frame of media for every line
+- Added /frame endpoint
 
 ## 2.0 (2025-12-30)
 **Changes**